Refactored directories and namespaces
