+++ Markdown/Markdown/Factories/Markdown/MarkdownTokenFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Markdown.Tokens;
 
-namespace Markdown.Factories
+namespace Markdown.Factories.Markdown
 {
     public class MarkdownTokenFactory : ITokenFactory<MarkdownToken>
     {
